+++ test/array.spec.js
@@ -131,15 +131,10 @@ describe('Array Functions', () => {
   })
   it('intersperse', () => {
     expect(
-      F.intersperse((prev, next) => next.length === 1 ? 'and finally' : 'and', [1, 2, 3])
+      F.intersperse((acc, i, xs) => i === xs.length - 1 ? 'and finally' : 'and', [1, 2, 3])
     ).to.deep.equal([ 1, 'and', 2, 'and finally', 3])
     expect(
       F.intersperse('and', [1, 2, 3])
     ).to.deep.equal([ 1, 'and', 2, 'and', 3])
-  })
-  it('intersperseGrammar', () => {
-    expect(
-      F.intersperseGrammar('or', 'or perhaps', ['first', 'second', 'third'])
-    ).to.deep.equal(['first', 'or', 'second', 'or perhaps', 'third'])
   })
 })